+++ test/runtests.jl
@@ -15,8 +15,8 @@ using ReferenceTests
         @info "Testing heatmaps..."
         include("test_heatmap.jl")
     end
-    @testset "XAIBase Explanations" begin
-        @info "Testing heatmaps on XAIBase explanations..."
+    @testset "XAIBase Attributions" begin
+        @info "Testing heatmaps on XAIBase attributions..."
         include("test_xaibase.jl")
     end
 end